requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.8
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-tenants==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
